Fix and addresses suggestions
